refactor: rename to follow conventions
